sw64: kvm: move struct kvm_regs to uapi kvm.h
