+++ arch/sw_64/include/asm/kvm.h
@@ -13,71 +13,6 @@
 #define SWVM_NUM_NUMA_MEMBANKS	1
 
 /*
- * for KVM_GET_REGS and KVM_SET_REGS
- */
-#ifdef CONFIG_SUBARCH_C3B
-struct kvm_regs {
-	unsigned long r0;
-	unsigned long r1;
-	unsigned long r2;
-	unsigned long r3;
-
-	unsigned long r4;
-	unsigned long r5;
-	unsigned long r6;
-	unsigned long r7;
-
-	unsigned long r8;
-	unsigned long r9;
-	unsigned long r10;
-	unsigned long r11;
-
-	unsigned long r12;
-	unsigned long r13;
-	unsigned long r14;
-	unsigned long r15;
-
-	unsigned long r19;
-	unsigned long r20;
-	unsigned long r21;
-	unsigned long r22;
-
-	unsigned long r23;
-	unsigned long r24;
-	unsigned long r25;
-	unsigned long r26;
-
-	unsigned long r27;
-	unsigned long r28;
-	unsigned long reserved;
-	unsigned long fpcr;
-
-	unsigned long fp[124];
-	/* These are saved by HMcode: */
-	unsigned long ps;
-	unsigned long pc;
-	unsigned long gp;
-	unsigned long r16;
-	unsigned long r17;
-	unsigned long r18;
-	unsigned long __padding[6];
-};
-
-#elif CONFIG_SUBARCH_C4
-struct kvm_regs {
-	union {
-		struct user_pt_regs regs;
-		struct {
-			unsigned long r[31];
-			unsigned long pc;
-			unsigned long ps;
-		};
-	};
-	struct user_fpsimd_state fpstate;
-};
-#endif
-
-/*
  * return stack for __sw64_vcpu_run
  */
 struct vcpu_run_ret_stack {
